refactor: moved the code from App.js into separate components <Game/> and <GameField/>
