+++ src/pages/Game.module.css
@@ -1,0 +1,6 @@
+.wrapper {
+  margin-top: 2%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+ }
